Feat/audit logs

+++ src/main/java/com/sifa/core_sifa/repository/IAuditLogRepository.java
@@ -8,7 +8,7 @@ import com.sifa.core_sifa.model.AuditLog;
 import java.util.List;
 
 @Repository
-public interface IAuditLogRepository extends JpaRepository<AuditLog, Integer> {
+public interface IAuditLogRepository extends JpaRepository<AuditLog, Long> {
 
     // traer todas las auditorias por el email del usuario
     List<AuditLog> findByEmailUsuario(String emailUsuario);
